moved remainder of ast_ephem into assist_spk_calc
