+++ unit_tests/benchmark/benchmark_rein.txt
@@ -65,3 +65,8 @@
 0.892357s 0.000578s 1.0.1b7 Jan 29 2023 21:44:05 34021eae3ff91ac7e178cc2bfd5be30f1d1a943c
 0.896800s 0.002069s 1.0.1b7 Jan 29 2023 21:44:05 34021eae3ff91ac7e178cc2bfd5be30f1d1a943c
 0.895708s 0.000922s 1.0.1b7 Jan 29 2023 21:44:05 34021eae3ff91ac7e178cc2bfd5be30f1d1a943c
+0.893079s 0.000503s 1.0.1b7 Jan 29 2023 22:06:19 aab08a04bcd6981c0cd34e61e43377d4eeb6ec08
+0.896796s 0.001000s 1.0.1b7 Jan 29 2023 22:06:19 aab08a04bcd6981c0cd34e61e43377d4eeb6ec08
+0.887684s 0.001910s 1.0.1b7 Jan 29 2023 22:19:26 aab08a04bcd6981c0cd34e61e43377d4eeb6ec08
+0.885797s 0.001569s 1.0.1b7 Jan 29 2023 22:19:26 aab08a04bcd6981c0cd34e61e43377d4eeb6ec08
+0.889591s 0.001523s 1.0.1b7 Jan 29 2023 22:19:26 aab08a04bcd6981c0cd34e61e43377d4eeb6ec08
